+++ packages/ui/src/hooks.server.ts
@@ -1,14 +1,19 @@
 import type {Handle} from '@sveltejs/kit'
 
 export const handle = (async ({event, resolve}) => {
-	const appSettings = await Promise.resolve(event.cookies.get('fat-fuzzy-settings-app'))
-	const navReveal = await Promise.resolve(event.cookies.get('fat-fuzzy-nav-reveal'))
-	const settingsReveal = await Promise.resolve(event.cookies.get('fat-fuzzy-settings-reveal'))
-	const sidebarReveal = await Promise.resolve(event.cookies.get('fat-fuzzy-sidebar-app'))
-	const styles = await Promise.resolve(event.cookies.get('fat-fuzzy-ui'))
+	const appSettings = event.cookies.get('fat-fuzzy-settings-app')
+	const navReveal = event.cookies.get('fat-fuzzy-nav-reveal')
+	const settingsReveal = event.cookies.get('fat-fuzzy-settings-reveal')
+	const sidebarReveal = event.cookies.get('fat-fuzzy-sidebar-reveal')
+	const dsContext = event.cookies.get('fat-fuzzy-ui-context-reveal')
+	const dsStyles = event.cookies.get('fat-fuzzy-ui-styles')
+	const dsState = event.cookies.get('fat-fuzzy-ui-state')
 
-	if (styles) {
-		event.locals.styles = styles
+	if (navReveal) {
+		event.locals.nav = navReveal
+	}
+	if (sidebarReveal) {
+		event.locals.sidebar = sidebarReveal
 	}
 	if (settingsReveal) {
 		event.locals.settings = settingsReveal
@@ -16,13 +21,15 @@ export const handle = (async ({event, resolve}) => {
 	if (appSettings) {
 		event.locals.app = appSettings
 	}
-	if (navReveal) {
-		event.locals.nav = navReveal
+	if (dsState) {
+		event.locals.dsState = dsState
 	}
-	if (sidebarReveal) {
-		event.locals.nav = sidebarReveal
+	if (dsStyles) {
+		event.locals.dsStyles = dsStyles
 	}
-
+	if (dsContext) {
+		event.locals.dsContext = dsContext
+	}
 	const response = await resolve(event)
 	return response
 }) satisfies Handle