+++ packages/ui/src/lib/components/blocks/buttons/Button.svelte
@@ -2,23 +2,29 @@
 	import type {ButtonType} from '$types'
 	import {createEventDispatcher} from 'svelte'
 
+	/**
+	 * State props
+	 */
 	export let id = 'button'
 	export let title = ''
 	export let name = ''
-	export let disabled = false
-	export let color = ''
-	export let variant = 'fill'
-	export let layout = 'switcher'
-	export let breakpoint = ''
-	export let size = ''
-	export let align = ''
-	export let asset = '' // emoji:value or svg:value
-	export let shape = ''
 	export let text = ''
 	export let value = ''
-	export let formaction: string | undefined = undefined
-	// export let pathname: string | undefined = undefined
+	export let disabled = false
 	export let type: ButtonType = 'submit'
+	export let formaction: string | undefined = undefined
+
+	/**
+	 * Style props
+	 */
+	export let align = ''
+	export let asset = '' // emoji:value or svg:value
+	export let breakpoint = ''
+	export let color = ''
+	export let layout = 'switcher'
+	export let size = ''
+	export let shape = ''
+	export let variant = 'fill'
 
 	const dispatch = createEventDispatcher()