+++ packages/ui/src/routes/doc/+layout.svelte
@@ -1,10 +1,15 @@
 <script lang="ts">
+	import type {LayoutData} from './$types'
 	import {onDestroy} from 'svelte'
 	import {page} from '$app/stores'
 
-	import * as settings from '$stores/settings'
+	import * as settingsStore from '$stores/settings'
+	import * as uiStore from '$stores/ui'
 
 	import {tokens, blocks, compositions, layouts} from '$lib'
+
+	export let data: LayoutData
+
 	const {RevealNav} = compositions
 	let path = $page.url.pathname
 
@@ -13,16 +18,22 @@
 		return a < b ? -1 : b < a ? 1 : 0
 	}
 
+	const {sidebar, styles, context} = data
+
+	settingsStore.sidebarReveal.set(sidebar)
+	uiStore.styles.set(styles)
+	uiStore.reveal.set(context)
+
 	const tokenNames = Object.keys(tokens).sort(sortAsc)
 	const blockNames = Object.keys(blocks).sort(sortAsc)
 	const layoutNames = Object.keys(layouts).sort(sortAsc)
 	const compositionNames = Object.keys(compositions).sort(sortAsc)
 	let title = 'Fat Fuzzy Test' // TODO : Fix title in children components: add breadcrumb nav component ?
 
-	let sidebarReveal: {[key: string]: string} = {reveal: ''}
+	let sidebarReveal: {[key: string]: string} = sidebar || {reveal: ''}
 
 	const stores = [
-		settings.sidebarReveal.subscribe((value) => {
+		settingsStore.sidebarReveal.subscribe((value) => {
 			if (value) {
 				sidebarReveal = value
 			}
@@ -76,6 +87,7 @@
 			size="md"
 			color="bg:primary:light"
 			position="fixed"
+			place="left"
 			background="polar"
 			container="card"
 			formaction="toggleSidebar"